+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Version 0.2.0
 - Add support for single quoted string literals
 - Remove i4gl sources from final image
 - Implement initial icgi library support
+- Add fgl_getenv builtin
 
 Version 0.1.0